requirements.txt
adal==1.2.2
azure-common==1.1.25
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-nspkg==3.0.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
Flask==1.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mock==4.0.2
mongolog==0.1.2
motor==2.1.0
msrest==0.6.13
msrestazure==0.6.3
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
python-memcached==1.59
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tornado==6.0.4
urllib3==1.25.8
Werkzeug==1.0.1
WTForms==2.2.1
